+++ framework/cognito_client.go
@@ -1,4 +1,4 @@
-// Copyright © 2016 Tom Maiaroto <tom@shift8creative.com>
+// Copyright © 2016 Tom Maiaroto <tom@SerifAndSemaphore.io>
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -51,7 +51,7 @@ type CognitoAppClient struct {
 	LogoutRedirectURI        string
 	TokenEndpoint            string
 	Base64BasicAuthorization string
-	Tracer                   TraceStrategy
+	Tracer                   XRayTraceStrategy
 }
 
 // CognitoAppClientConfig defines required info to build a new CognitoAppClient