+++ framework/s3.go
@@ -1,4 +1,4 @@
-// Copyright © 2016 Tom Maiaroto <tom@shift8creative.com>
+// Copyright © 2016 Tom Maiaroto <tom@SerifAndSemaphore.io>
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,7 +17,6 @@ package framework
 import (
 	"context"
 	"errors"
-	"log"
 
 	"github.com/aws/aws-lambda-go/lambda"
 	"github.com/gobwas/glob"
@@ -50,7 +49,6 @@ func (r *S3ObjectRouter) LambdaHandler(ctx context.Context, d *HandlerDependenci
 		return errors.New("no handlers registered for S3ObjectRouter")
 	}
 	var err error
-	var g glob.Glob
 	handled := false
 
 	// There can apparently be multiple, a router handler is to handle one object/file operation
@@ -61,27 +59,33 @@ func (r *S3ObjectRouter) LambdaHandler(ctx context.Context, d *HandlerDependenci
 		if r.handlers != nil {
 			// and if the bucket matches or if no bucket was defined for the S3ObjectRouter (all buckets)
 			if r.Bucket == "" || r.Bucket == record.S3.Bucket.Name {
+
 				// Each key on handlers map is a string that can be used to glob match
 				// If one of them matches, use that
 				for globStr, handler := range r.handlers {
-					g = glob.MustCompile(globStr)
-					if g.Match(record.S3.Object.Key) {
-						handled = true
-						// Trace (default is to use XRay)
-						// Annotations can be searched in XRay.
-						r.Tracer.Annotations = map[string]interface{}{
-							"S3Bucket":    record.S3.Bucket.Name,
-							"S3ObjectKey": record.S3.Object.Key,
-							"S3Event":     record.EventName,
+					// and also if the event matches or if no event was defined for the S3ObjectHandler (all event sources)
+					// and of course this also matches by glob match too with values like `s3:ObjectCreated:*` and `s3:ObjectCreated:Put`
+					esG, err := glob.Compile(handler.Event)
+					if handler.Event == "" || (err == nil && esG.Match(record.EventSource)) {
+						g, err := glob.Compile(globStr)
+						if err == nil && g.Match(record.S3.Object.Key) {
+							handled = true
+							// NOTE: If Tracer is nil, we have a problem.
+							r.Tracer.Record("annotation",
+								map[string]interface{}{
+									"S3Bucket":    record.S3.Bucket.Name,
+									"S3ObjectKey": record.S3.Object.Key,
+									"S3Event":     record.EventName,
+								},
+							)
+							err = r.Tracer.Capture(ctx, "S3ObjectHandler", func(ctx1 context.Context) error {
+								d.Tracer = &r.Tracer
+								return handler.Handler(ctx1, d, &evt)
+							})
 						}
-						err = r.Tracer.Capture(ctx, "S3ObjectHandler", func(ctx1 context.Context) error {
-							r.Tracer.AddAnnotations(ctx1)
-							r.Tracer.AddMetadata(ctx1)
-							d.Tracer = &r.Tracer
-							return handler.Handler(ctx1, d, &evt)
-						})
 					}
 					// TODO: think about some verbose setting for the framework.
+					// Possibly trace an error
 					// else {
 					// 	log.Println("glob no match")
 					// }
@@ -91,19 +95,20 @@ func (r *S3ObjectRouter) LambdaHandler(ctx context.Context, d *HandlerDependenci
 				// The application can inspect the map and make a decision on what to do, if anything.
 				// This is optional.
 				if !handled {
-					log.Println("using default fall through handler")
 					// It's possible that the S3ObjectRouter wasn't created with NewS3ObjectRouter, so check for this still.
-					if handler, ok := r.handlers["*"]; ok {
-						// Capture the handler (in XRay by default) automatically
-						r.Tracer.Annotations = map[string]interface{}{
-							"S3Bucket":           record.S3.Bucket.Name,
-							"S3ObjectKey":        record.S3.Object.Key,
-							"S3Event":            record.EventName,
-							"FallthroughHandler": true,
-						}
+					if handler, ok := r.handlers["_"]; ok {
+						r.Tracer.Record("annotation",
+							map[string]interface{}{
+								"S3Bucket":           record.S3.Bucket.Name,
+								"S3ObjectKey":        record.S3.Object.Key,
+								"S3Event":            record.EventName,
+								"FallthroughHandler": true,
+							},
+						)
+
 						err = r.Tracer.Capture(ctx, "S3ObjectHandler", func(ctx1 context.Context) error {
-							r.Tracer.AddAnnotations(ctx1)
-							r.Tracer.AddMetadata(ctx1)
+							// Tracer is passed in as a dependency so that the handler (user function) can also
+							// add annotations, metadata, etc. Whatever the TraceStrategy allows for.
 							d.Tracer = &r.Tracer
 							return handler.Handler(ctx1, d, &evt)
 						})
@@ -125,17 +130,18 @@ func (r *S3ObjectRouter) Listen() {
 // NewS3ObjectRouter simply returns a new S3ObjectRouter struct and behaves a bit like Router, it even takes an optional rootHandler or "fall through" catch all
 func NewS3ObjectRouter(rootHandler ...S3ObjectHandler) *S3ObjectRouter {
 	// The catch all is optional, if not provided, an empty handler is still called and it returns nothing.
+	// Note: No Event field, so all event sources match for the fall through
 	handler := S3ObjectHandler{
 		Handler: func(context.Context, *HandlerDependencies, *S3Event) error {
 			return nil
 		},
 	}
-	if len(rootHandler) > 0 {
+	if rootHandler != nil {
 		handler = rootHandler[0]
 	}
 	return &S3ObjectRouter{
 		handlers: map[string]S3ObjectHandler{
-			"*": handler,
+			"_": handler,
 		},
 	}
 }
@@ -143,7 +149,7 @@ func NewS3ObjectRouter(rootHandler ...S3ObjectHandler) *S3ObjectRouter {
 // NewS3ObjectRouterForBucket is the same as NewS3ObjectRouter except it's for a specific bucket (you could also set the bucket field after using the other function)
 func NewS3ObjectRouterForBucket(bucket string, rootHandler ...S3ObjectHandler) *S3ObjectRouter {
 	var r *S3ObjectRouter
-	if len(rootHandler) > 0 {
+	if rootHandler != nil {
 		r = NewS3ObjectRouter(rootHandler[0])
 	} else {
 		r = NewS3ObjectRouter()